refactor variant
